Module 11

+++ module11/Convolution.cl
@@ -14,8 +14,8 @@
 
 __kernel void convolve(
 	const __global  uint * const input,
-    __constant uint * const mask,
-    __global  uint * const output,
+    __constant float * const mask,
+    __global  float * const output,
     const int inputWidth,
     const int maskWidth)
 {
